Create a page for the editor.
