+++ src/App.jsx
@@ -1,0 +1,7 @@
+import React, { useState } from 'react';
+import Editor from './components/Editor';
+
+function App() {
+}
+
+export default App;
